+++ mnist.h
@@ -4,6 +4,7 @@
 #define ROWS 1000
 #define COLS 28 * 28
 
-void load_csv(const char *filename, double data[ROWS][COLS], int labels[ROWS]);
+void load_csv(const char *filename, double data[ROWS][COLS],
+              double labels[ROWS]);
 
 #endif // MNIST_H
